Moved all components to the core namespace
Moved all components to the core namespace, making it more simple in the future to create themes, upgrades, and extensions.
