+++ src/components/Header.js
@@ -1,9 +1,9 @@
 import Link from 'next/link';
 import { useRouter } from 'next/router';
-import VersionSwitcher from '@/components/Header/VersionSwitcher';
+import VersionSwitcher from '@/components/core/Header/VersionSwitcher';
 import { SearchButton } from '@/components/Search';
 import Router from 'next/router';
-import Logo from '@/components/Logos/Logo';
+import Logo from '@/components/core/Logos/Logo';
 import { Dialog } from '@headlessui/react';
 import { useEffect, useState } from 'react';
 import clsx from 'clsx'
