+++ src/components/core/NewsletterForm/NewsletterForm.test.tsx
