+++ src/pages/docs/deployment.mdx
@@ -3,8 +3,8 @@ title: Deployment
 description: A guide to deploying your Elegant application to production.
 ---
 
-import DeploymentSteps from '@/components/Docs/DeploymentSteps';
-import CustomDomainSteps from '@/components/Docs/CustomDomainSteps';
+import DeploymentSteps from '@/components/core/Docs/DeploymentSteps';
+import CustomDomainSteps from '@/components/core/Docs/CustomDomainSteps';
 import vercel_homepage from '@/img/docs/Vercel.com-Homepage.png'
 
 - [Introduction](#introduction)
