+++ src/pages/index.js
@@ -2,9 +2,9 @@ import React from 'react';
 import Image from 'next/image';
 import { CodeWindow } from '@/components/CodeWindow';
 import { SearchButton } from '@/components/Search';
-import { NewsletterForm } from '@/components/NewsletterForm/NewsletterForm';
+import { NewsletterForm } from '@/components/core/NewsletterForm/NewsletterForm';
 import UnifiedFramework from '@/components/home/UnifiedFramework';
-import Logo from '@/components/Logos/Logo';
+import Logo from '@/components/core/Logos/Logo';
 import { Footer } from '@/components/home/Footer';
 import NextLink from 'next/link';
 import Head from 'next/head';
